refactor(cron): extract JobEditDialog with accordion sections
